add 'local' flag, move to 'io'
